requirements.txt
transformers
fastapi
uvicorn
torch
python-dotenv